refactor: move code + update readme
