+++ README.md
@@ -3,7 +3,7 @@ Starter for the HiFive RevB board by SiFive.
 
 ### TODO
 
-- [ ] Implement with DIM LEDS
+- [X] Implement with DIM LEDS
 - [X] Add LED COLOR CYCLE
 - [X] Test on hardware
 - [X] Add remaining addr/offset defines
