+++ src/_extra/platform.h
